add missing newline to last string in StringBuilder array
This is a one line change to keep lint from complaining about missing linefeeds in sandbox.ts.  This change does not handle the case where the StringBuilder array is empty when dump is called.  For the use case this doesn't seem like an issue.  One could add a check for data being an empty string (null check is not required as join is "guaranteed" to return a valid string.