improve performance of visual regressions build
this was done by exposing a function on the window to load a new scenario, instead of navigating to a new url which required reloading all the page resources.

also added new options for the visual regressions build, a list of excluded paths (by regex), and the mock date to set for the visual regression checks (useful for components that use the current date).